requirements.txt
python-telegram-bot==20.7
pyyaml==6.0.1
psycopg2-binary==2.9.9
requests==2.31.0
httpx>=0.27.0
redis>=5.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
